Some fixes.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@31432 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/arch/beos/hs.h
@@ -35,8 +35,8 @@ extern int hs_pci_open(void);
 extern int hs_isa_close(void);
 extern int hs_pci_close(void);
 
-extern int hs_isa_read(WORD addr, int chipno);
-extern int hs_pci_read(WORD addr, int chipno);
+extern BYTE hs_isa_read(WORD addr, int chipno);
+extern BYTE hs_pci_read(WORD addr, int chipno);
 
 extern void hs_isa_store(WORD addr, BYTE val, int chipno);
 extern void hs_pci_store(WORD addr, BYTE val, int chipno);
